refactor(product): Category, ProductCondition, InspectionStatus 등 enum클래스들을 shared 경로로 이동
