+++ src/main/java/com/bugzero/rarego/boundedContext/product/app/ProductApiAdapter.java
@@ -10,7 +10,6 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.Product;
 import com.bugzero.rarego.boundedContext.product.domain.ProductImage;
 import com.bugzero.rarego.boundedContext.product.out.InspectionRepository;
@@ -18,6 +17,7 @@ import com.bugzero.rarego.boundedContext.product.out.ProductImageRepository;
 import com.bugzero.rarego.boundedContext.product.out.ProductRepository;
 import com.bugzero.rarego.shared.product.dto.ProductAuctionResponseDto;
 import com.bugzero.rarego.shared.product.out.ProductApiClient;
+import com.bugzero.rarego.shared.product.type.InspectionStatus;
 
 import lombok.RequiredArgsConstructor;
