+++ src/main/java/com/bugzero/rarego/boundedContext/product/domain/dto/ProductInspectionRequestDto.java
@@ -1,10 +1,10 @@
 package com.bugzero.rarego.boundedContext.product.domain.dto;
 
 import com.bugzero.rarego.boundedContext.product.domain.Inspection;
-import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.Product;
-import com.bugzero.rarego.boundedContext.product.domain.ProductCondition;
 import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
+import com.bugzero.rarego.shared.product.type.InspectionStatus;
+import com.bugzero.rarego.shared.product.type.ProductCondition;
 
 import jakarta.validation.constraints.NotNull;
