+++ src/main/java/com/bugzero/rarego/boundedContext/product/domain/dto/ProductInspectionResponseDto.java
@@ -2,8 +2,8 @@ package com.bugzero.rarego.boundedContext.product.domain.dto;
 
 import java.time.LocalDateTime;
 
-import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
-import com.bugzero.rarego.boundedContext.product.domain.ProductCondition;
+import com.bugzero.rarego.shared.product.type.InspectionStatus;
+import com.bugzero.rarego.shared.product.type.ProductCondition;
 
 import lombok.Builder;
