+++ src/main/java/com/bugzero/rarego/boundedContext/product/domain/dto/ProductResponseForInspectionDto.java
@@ -1,7 +1,7 @@
 package com.bugzero.rarego.boundedContext.product.domain.dto;
 
-import com.bugzero.rarego.boundedContext.product.domain.Category;
-import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
+import com.bugzero.rarego.shared.product.type.Category;
+import com.bugzero.rarego.shared.product.type.InspectionStatus;
 
 import lombok.Builder;
