+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductCreateRequestDto.java
@@ -3,12 +3,12 @@ package com.bugzero.rarego.shared.product.dto;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.bugzero.rarego.boundedContext.product.domain.Category;
-import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.Product;
-import com.bugzero.rarego.boundedContext.product.domain.ProductCondition;
 import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
 import com.bugzero.rarego.boundedContext.product.domain.dto.ProductImageRequestDto;
+import com.bugzero.rarego.shared.product.type.Category;
+import com.bugzero.rarego.shared.product.type.InspectionStatus;
+import com.bugzero.rarego.shared.product.type.ProductCondition;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
